Solved day 7, moved day interface from int to long as the answers are getting to big :+)

+++ src/main/java/nl/th8/adventofcode2024/Day.java
@@ -2,9 +2,9 @@ package nl.th8.adventofcode2024;
 
 public interface Day {
 
-    int solvePartOne();
+    long solvePartOne();
 
-    int solvePartTwo();
+    long solvePartTwo();
 
     int getDayNumber();
 }
